src/mplayer.c:   - don't start a new mplayer if an old mplayer is already running
src/mplayer.c, src/keyfile.c, src/keyfile.h:
  - fix build warnings

+++ src/keyfile.c
@@ -42,7 +42,7 @@ gboolean key_file_has_group(const gchar *group_name)
 }
 
 /* -------------------------------------------------------------------------- */
-gchar **key_file_get_keys(const gchar *group_name, int *number)
+gchar **key_file_get_keys(const gchar *group_name, gsize *number)
 {
     return g_key_file_get_keys(s_key_file, group_name, number, NULL);
 }
